V2/develop frontend

+++ DataAcces/Repositories/CommentRepository.cs
@@ -13,30 +13,48 @@ namespace DataAcces.Repositories
         {
             _context = context;
         }
+
         public IEnumerable<Comment> GetAllComments()
         {
             return _context.Comments.ToList();
         }
+
         public Comment GetCommentById(int id)
         {
             return _context.Comments.Find(id);
         }
+
         public void InsertComment(Comment comment)
         {
             _context.Comments.Add(comment);
         }
+
         public void UpdateComment(Comment comment)
         {
             _context.Entry(comment).State = System.Data.Entity.EntityState.Modified;
         }
+
         public void DeleteComment(int id)
         {
-            Comment comment = _context.Comments.Find(id);
-            _context.Comments.Remove(comment);
+            var comment = _context.Comments.Find(id);
+            if (comment != null)
+            {
+                _context.Comments.Remove(comment);
+            }
         }
+
         public void DeleteAll()
         {
-            _context.Comments.RemoveRange(_context.Comments);
+            var comments = _context.Comments.ToList();
+            _context.Comments.RemoveRange(comments);
+        }
+
+        public List<Comment> GetCommentsByTicketId(int ticketId)
+        {
+            return _context.Comments
+                .Where(c => c.TicketID == ticketId)
+                .OrderBy(c => c.PostedAt)
+                .ToList();
         }
     }
 }
